+++ docs/changelog.md
@@ -5,9 +5,10 @@
 ### Common changes
 
 - Added `not:` prefix for negate result of checked condition (can be used in hooks and offers and in both programatic and content pack usage)
-- Added new condition `Random` with random chance in % to offer or execute hook
 - Added new conditions for check known recipes: `KnownCraftingRecipe` and `KnownCookingRecipe`
-- Added new conditions for check if quest was current season, day, year and other time periods: `QuestAcceptedInTimePeriod`
+- Added new condition `Random` with random chance in % to offer or execute hook
+- Added new condition `QuestAcceptedDate` for check if quest was accepted in specified date
+- Added new condition `QuestAcceptedInPeriod` for check if quest was accepted in current season, day, year and other time period(s).
 - Added stats support
 - Improved some minor refactors